+++ src/components/domains/myPage/contentSection/study/FavoriteStudyList/index.tsx
