+++ src/components/domains/myPage/contentSection/study/OpenedStudyAdministration/index.tsx
