+++ src/components/domains/myPage/contentSection/study/ParticipatingStudy/index.tsx
