+++ src/components/domains/myPage/contentSection/study/StudyApplicantConfirmation/index.tsx
