+++ src/pages/MyPage/index.scss
@@ -3,7 +3,7 @@
   background-color: #F9F9F9;
   margin-top: 70px;
 
-  .myPageContentSection {
+  .contentSection {
     width: 100%;
     padding: 0 50px;
   }
