+++ src/pages/MyPage/index.tsx
@@ -1,8 +1,8 @@
+import UserProfile from '@components/domains/myPage/common/UserProfile';
+import { defaultFocusMenu } from '@components/domains/myPage/MyPageMenu';
+import SideBar from '@components/domains/myPage/Sidebar';
 import React, { useState } from 'react';
 import './index.scss';
-import SideBar from '@components/domains/myPage/sidebar';
-import UserProfile from '@components/domains/myPage/components/common/UserProfile';
-import { defaultFocusMenu } from '@components/domains/myPage/MyPageMenu';
 
 const MyPage = () => {
   const [focusedMenu, setFocusedMenu] = useState<any>(defaultFocusMenu);
@@ -10,7 +10,7 @@ const MyPage = () => {
   return (
     <div className="myPage">
       <SideBar focusedMenu={focusedMenu} setFocusedMenu={setFocusedMenu} />
-      <section className="myPageContentSection">
+      <section className="contentSection">
         <UserProfile />
         {focusedMenu.component}
       </section>
